Password Generator
This password generator can generate and save generated passwords to your computer. This tool is very useful for those who have trouble remembering or creating passwords. Why use your head when your computer can do it for you? :). Send feedbacks to @EyetuKingsley on twitter and also check twitter for the video I made for the tool.